requirements.txt
django
coverage
djangorestframework
django-cors-headers
redis
rq
json